Refactoring to make ext data accessiable as common functionality
