+++ src/Tug.Base/Util/ExtDataExtensions.cs
@@ -1,3 +1,8 @@
+/*
+ * Copyright © The DevOps Collective, Inc. All rights reserved.
+ * Licnesed under GNU GPL v3. See top-level LICENSE.txt for more details.
+ */
+
 using System.Collections.Generic;
 using Newtonsoft.Json.Linq;
